# uStoreAPI/uStoreAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "uStoreAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Horarios/GetHorarios": {
      "get": {
        "tags": [
          "Horarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/HorarioDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Horarios": {
      "get": {
        "tags": [
          "Horarios"
        ],
        "operationId": "GetHorario",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HorarioDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Horarios/CreateHorario": {
      "post": {
        "tags": [
          "Horarios"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/HorarioDto"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/HorarioDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Horarios/DeleteHorarios": {
      "delete": {
        "tags": [
          "Horarios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Horarios/UpdateHorario": {
      "put": {
        "tags": [
          "Horarios"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/HorarioDto"
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "HorarioDto": {
      "required": [
        "dia",
        "horarioApertura",
        "horarioCierre",
        "idTienda"
      ],
      "type": "object",
      "properties": {
        "idHorario": {
          "format": "int32",
          "type": "integer"
        },
        "dia": {
          "minLength": 1,
          "type": "string"
        },
        "horarioApertura": {
          "minLength": 1,
          "type": "string"
        },
        "horarioCierre": {
          "minLength": 1,
          "type": "string"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
  }
}